rename utils\validate to schemas\contactsSchema
