+++ controllers/contacts/updateContact.js
@@ -1,7 +1,7 @@
 const { isValidObjectId } = require('mongoose');
 const { contactsService } = require('../../service/');
 
-const { schemaUpdate } = require('../../utils/validate');
+const { schemaUpdate } = require('../../schemas/contactsSchema');
 const { httpError } = require('../../helpers');
 
 const updateContact = async (req, res, next) => {
