+++ controllers/contacts/updateFavorite.js
@@ -1,5 +1,5 @@
 const { isValidObjectId } = require('mongoose');
-const { schemaUpdate } = require('../../utils/validate');
+const { schemaUpdate } = require('../../schemas/contactsSchema');
 const { contactsService } = require('../../service');
 const { httpError } = require('../../helpers');
