+++ middlewares/handleMangooseError.js
@@ -1,0 +1,6 @@
+const handleMongooseError = (error, data, next) => {
+  error.status = 400;
+  next();
+};
+
+module.exports = handleMongooseError;
