+++ schemas/index.js
@@ -1,0 +1,3 @@
+const contactsSchema = require('./contactsSchema');
+
+module.exports = { contactsSchema };
